+++ packages/moonstone/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
+## [unreleased]
+
+### Changed
+
+- `moonstone/Scroller` to scroll when no spottable child exists in the 5-way key direction then focus a scrollbar button if `focusableScrollbar` is true
+
 ## [3.0.0-alpha.2] - 2019-05-20
 
 ### Added